LayoutTests/platform/ios/fast/events/ios/rotation/resize-iframe-after-orientation-change-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById('frame').scrollWidth is outer.scrollWidth
FAIL document.getElementById('frame').contentDocument.getElementById('inner').scrollWidth should be 800. Was 390.
Each time the viewport is resized, the outer div (green), iframe (orange) and inner div (pink) should take the full viewport width:
